NewsTrader/Scripts/requirements.txt
alpaca-py
httpx
langchain-anthropic
langchain-core
langchain-ollama
langchain-openai
openpyxl
pandas
pydantic>=2
python-dotenv
pytz
reportlab
requests
tiingo
